backend/requirements.txt
# Backend Dependencies for PrepAI
# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
pydantic==2.12.3
pydantic_core==2.41.4
starlette==0.37.2

# AI Integration
google-generativeai==0.8.5
google-genai==1.46.0
google-auth==2.41.1
google-api-core==2.27.0
grpcio==1.76.0

# Database
pymongo==4.5.0
motor==3.3.1

# PDF Processing
pdf2image==1.17.0
PyMuPDF==1.23.14
PyPDF2==3.0.1
pillow==12.0.0

# Authentication & Security
passlib==1.7.4
python-jose==3.5.0
bcrypt==4.0.1
PyJWT==2.10.1

# Environment & Forms
python-dotenv==1.1.1
python-multipart==0.0.20

# Data Processing
numpy==1.26.4
pandas==1.5.3

# HTTP & Async
httpx==0.28.1
aiohttp==3.13.1

# Email Validation
email-validator==2.3.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2